api/requirements.txt
fastapi
uvicorn
python-multipart
manim
python-dotenv
